moved planner and workouthistory
