+++ frontend/src/components/program/components/index.ts
@@ -1,5 +1,6 @@
 import { AddProgram } from './AddProgram'
 import { ProgramDetails } from './ProgramDetails'
 import { ProgramList } from './ProgramList'
+import { Planner } from './Planner'
 
-export { AddProgram, ProgramDetails, ProgramList }
+export { AddProgram, ProgramDetails, ProgramList, Planner }
