+++ frontend/src/components/workout/components/index.ts
@@ -1,5 +1,6 @@
 import { WorkoutHistoryTableRow } from './WorkoutHistoryTableRow'
 import { WorkoutDetails } from './WorkoutDetails'
 import { WorkoutList } from './WorkoutList'
+import { WorkoutHistory } from './WorkoutHistory'
 
-export { WorkoutDetails, WorkoutHistoryTableRow, WorkoutList }
+export { WorkoutDetails, WorkoutHistoryTableRow, WorkoutList, WorkoutHistory }
